Complete Setup
To use the app, open a terminal to the base directory and run
`npm run install`. Then to start the server in one terminal run
`npm run frontend` and in another run `npm run backend`.

If it works right from this commit you should see a "hello from express"
below the welcome to react text.