Fixed SSL Token Authorization
Basically, we were trying to use HTTPS to send requests, and Vite
had the backend setup over HTTPS by default, but we didn't have any
certificates and so the browser was throwing an error. From now on,
lets plan to just use http.

In fact when we use `fetch` we can just omit the `http://localhost/` as
I modified them to be.

In addition to this, I removed some static files (db.sqlite and .env)
and included them in the .gitignore. In the future, we should include
instructions in the setup portion of the README on how to setup the
database and the .env file.